Fix orientation to portrait.
Ideally, there would be a landscape mode, and the bands would scroll
vertically through their colours. But as Band is derived from Gallery
and Gallery cannot scroll vertically, this can't happen yet.

Maybe one day Gallery can be extended or Band can be based on something
else, but in the mean time it's easiest to just disable landscape.