Add a Tolerance Band.
Resistance now has a tolerance member, with a setter and getter.
toString () now outputs its value.

New class ToleranceBand, similar to DigitBand and MultiplierBand.
processResistance () sets the resistance's tolerance.

Add a ToleranceBand to the layout below the multiplier band. Listen for
its colour changing and process it when updating the resistance.