Add Resistor class - a LinearLayout which contains multiple Bands.
This moves some of the code from the main Activity in to the new class.
The bands of a resistor are declared in xml as they were before.
The resistor is responsible for listening to its bands changing value
and updating its resistance. The resistance can be queried with
getResistance ().

There is not yet a listener for the resistance, meaning that the
program wont currently update the displayed value. Also note that
while the resistor is blue, the layout containing it and the text
is black. The text colour is also black, so cannot be seen.